requirements/dev.txt
-r base.txt
ipython==8.18.1
psycopg-binary==3.1.19
django-extensions==3.2.3
Faker==27.0.0 